backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
sentence-transformers
pypdf
python-dotenv
numpy
torch
scikit-learn
